Extend display of the task in list.
* Add coloring for task statuses.
* Add button to see task details.
* Return task status as enum from API.

Work on #8

+++ bff/src/Application/Projects/ProjectionWorkers/TaskStatusChangedProjectionWorkerBase.cs
@@ -33,7 +33,7 @@ public abstract class TaskStatusChangedProjectionWorkerBase<TAggregateEvent>(
         }
 
         var existingIndex = currentTasks.IndexOf(existing);
-        existing = existing with { Status = (int)this.TargetStatus };
+        existing = existing with { Status = this.TargetStatus };
         currentTasks[existingIndex] = existing;
         project = project with { Tasks = currentTasks };
         await projectionWriter.Update(project, cancellationToken);
